move instrumentation prompt funcs into subpackage
