+++ pkg/skaffold/instrumentation/export.go
@@ -45,7 +45,7 @@ import (
 )
 
 func ExportMetrics(exitCode int) error {
-	if !shouldExportMetrics || meter.Command == "" {
+	if !ShouldExportMetrics || meter.Command == "" {
 		return nil
 	}
 	home, err := homedir.Dir()
